Disentangle ScaledTimerImpl from TimerImpl
No need to use the actual TimerImpl in ScaledTimerImpl since it can get
by with only on the public interface methods.
