+++ source/common/event/timer_impl.cc
@@ -8,17 +8,6 @@
 
 namespace Envoy {
 namespace Event {
-namespace {
-
-void resetTimer(TimerImpl& timer, std::chrono::milliseconds d) {
-  timer.enableTimer(d, timer.scope());
-}
-
-void resetTimer(TimerImpl& timer, std::chrono::microseconds d) {
-  timer.enableHRTimer(d, timer.scope());
-}
-
-} // namespace
 
 TimerImpl::TimerImpl(Libevent::BasePtr& libevent, TimerCb cb, Dispatcher& dispatcher)
     : cb_(cb), dispatcher_(dispatcher) {
@@ -63,59 +52,6 @@ void TimerImpl::internalEnableTimer(const timeval& tv, const ScopeTrackedObject*
 }
 
 bool TimerImpl::enabled() { return 0 != evtimer_pending(&raw_event_, nullptr); }
-
-ScaledTimerImpl::ScaledTimerImpl(Libevent::BasePtr& libevent, TimerCb cb,
-                                 Event::Dispatcher& dispatcher, double scale_factor)
-    : scale_factor_(scale_factor > 1 ? 1 : !(scale_factor > 0) ? 0 : scale_factor),
-      timer_(libevent, cb, dispatcher) {}
-
-void ScaledTimerImpl::enableTimer(const std::chrono::milliseconds& min,
-                                  const std::chrono::milliseconds& max,
-                                  const ScopeTrackedObject* scope) {
-  ASSERT(max >= min);
-  enabled_.last_enabled = timer_.dispatcher().timeSource().monotonicTime();
-  enabled_.interval = Enabled::Interval<std::chrono::milliseconds>(min, max);
-  timer_.enableTimer(
-      std::chrono::duration_cast<std::chrono::milliseconds>(scale_factor_ * (max - min)) + min,
-      scope);
-}
-
-void ScaledTimerImpl::enableHRTimer(const std::chrono::microseconds& min,
-                                    const std::chrono::microseconds& max,
-                                    const ScopeTrackedObject* scope) {
-  ASSERT(max >= min);
-  enabled_.last_enabled = timer_.dispatcher().timeSource().monotonicTime();
-  enabled_.interval = Enabled::Interval<std::chrono::microseconds>(min, max);
-  timer_.enableHRTimer(
-      std::chrono::duration_cast<std::chrono::microseconds>(scale_factor_ * (max - min)) + min,
-      scope);
-}
-
-void ScaledTimerImpl::disableTimer() { timer_.disableTimer(); }
-
-bool ScaledTimerImpl::enabled() { return timer_.enabled(); }
-
-void ScaledTimerImpl::setScaleFactor(double scale_factor) {
-  scale_factor_ = scale_factor > 1 ? 1 : !(scale_factor > 0) ? 0 : scale_factor;
-
-  if (!timer_.enabled()) {
-    return;
-  }
-
-  auto visitor = [this](auto& interval) -> void {
-    using T = typename std::remove_reference_t<decltype(interval)>::value_type;
-
-    const auto trigger_time =
-        std::chrono::duration_cast<T>(scale_factor_ * (interval.max - interval.min)) +
-        interval.min + enabled_.last_enabled;
-
-    const T delay = std::chrono::duration_cast<T>(trigger_time -
-                                                  timer_.dispatcher().timeSource().monotonicTime());
-    resetTimer(timer_, std::max(T::zero(), delay));
-  };
-
-  absl::visit(visitor, enabled_.interval);
-}
 
 } // namespace Event
 } // namespace Envoy
